small changes and renamings
