+++ mesmerize_viz/baseviewer.py
